pyproject.toml
[project]
name = "langgraph-drafter-agent"
version = "0.1.0"
description = "AI based agent to draft anything"
authors = [
    {name = "tsamanta"}
]
readme = "README.md"
requires-python = "3.12.4"
dependencies = [
    "langgraph (>=0.5.2,<0.6.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-openai (>=0.3.27,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
